+++ AILZ80ASM.Test/AIEncodeTest.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using AILZ80ASM.AILight;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.IO;
