+++ AILZ80ASM.Test/AINameTest.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using AILZ80ASM.AILight;
+using AILZ80ASM.Assembler;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 
 namespace AILZ80ASM.Test
@@ -81,6 +83,24 @@ namespace AILZ80ASM.Test
             var arguments = AIName.ParseArguments("\"0123456789:;<=>? \"");
 
             Assert.AreEqual(arguments[0], "\"0123456789:;<=>? \"");
+        }
+
+        [TestMethod]
+        public void ReservedWordTest()
+        {
+            var reservedWords = new[] { "ORG" };
+            var asmLoad = new AsmLoad(new InstructionSet.Z80());
+
+            foreach (var reservedWord in reservedWords)
+            {
+                Assert.IsFalse(AIName.ValidateCharMapName(reservedWord, asmLoad));
+                Assert.IsFalse(AIName.ValidateFunctionName(reservedWord, asmLoad));
+                Assert.IsFalse(AIName.ValidateFunctionArgument(reservedWord, asmLoad));
+                Assert.IsFalse(AIName.ValidateMacroName(reservedWord, asmLoad));
+                Assert.IsFalse(AIName.ValidateMacroArgument(reservedWord, asmLoad));
+                Assert.IsFalse(AIName.DeclareLabelValidate(reservedWord, asmLoad));
+            }
+
         }
     }
 }
