+++ AILZ80ASM/AILight/AIName.cs
@@ -1,9 +1,10 @@
-﻿using System;
+﻿using AILZ80ASM.Assembler;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace AILZ80ASM
+namespace AILZ80ASM.AILight
 {
     public static class AIName
     {
@@ -55,11 +56,15 @@ namespace AILZ80ASM
         public static bool ValidateMacroName(string target, AsmLoad asmLoad)
         {
             if (string.IsNullOrEmpty(target))
+            {
                 return false;
+            }
 
             // 先頭に()は使えない
             if (target.StartsWith("(") || target.StartsWith(")"))
+            {
                 return false;
+            }
 
             return ValidateNameForMacroName(target, asmLoad);
         }
@@ -72,7 +77,9 @@ namespace AILZ80ASM
         public static bool ValidateMacroArgument(string target, AsmLoad asmLoad)
         {
             if (string.IsNullOrEmpty(target))
+            {
                 return false;
+            }
 
             return ValidateName(target, asmLoad);
         }
@@ -80,11 +87,15 @@ namespace AILZ80ASM
         public static bool ValidateFunctionName(string target, AsmLoad asmLoad)
         {
             if (string.IsNullOrEmpty(target))
+            {
                 return false;
+            }
 
             // ()は使えない
             if (target.IndexOfAny(new[] { '(', ')' }) != -1)
+            {
                 return false;
+            }
 
             return ValidateNameForFunction(target, asmLoad);
         }
@@ -97,7 +108,9 @@ namespace AILZ80ASM
         public static bool ValidateFunctionArgument(string target, AsmLoad asmLoad)
         {
             if (string.IsNullOrEmpty(target))
+            {
                 return false;
+            }
 
             return ValidateName(target, asmLoad);
         }
@@ -187,10 +200,12 @@ namespace AILZ80ASM
         /// </summary>
         /// <param name="target"></param>
         /// <returns></returns>
-        public static bool ValidateCharMapName(string target)
+        public static bool ValidateCharMapName(string target, AsmLoad asmLoad)
         {
-            if (string.IsNullOrEmpty(target))
+            if (!ValidateName(target, asmLoad))
+            {
                 return false;
+            }
 
             return Regex.Match(target, RegexPatternCharMapInvalid, RegexOptions.Singleline | RegexOptions.IgnoreCase).Success;
         }
@@ -230,7 +245,7 @@ namespace AILZ80ASM
             }
 
             return  Regex.Match(target, RegexPatternLabelValidate, RegexOptions.Singleline | RegexOptions.IgnoreCase).Success &&
-                   !Regex.Match(target, RegexPatternLabelInvalid, RegexOptions.Singleline | RegexOptions.IgnoreCase).Success;
+                   !Regex.Match(target, RegexPatternLabelInvalid,  RegexOptions.Singleline | RegexOptions.IgnoreCase).Success;
         }
 
         /// <summary>
