+++ AILZ80ASM/Assembler/AsmISA.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AILZ80ASM
+namespace AILZ80ASM.Assembler
 {
     public enum AsmISA
     {
