+++ AILZ80ASM/Assembler/AsmList.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
 
-namespace AILZ80ASM
+namespace AILZ80ASM.Assembler
 {
     public class AsmList
     {
