+++ AILZ80ASM/Function.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿using AILZ80ASM.AILight;
+using AILZ80ASM.Assembler;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
