+++ AILZ80ASM/InstructionSet/Z80.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using AILZ80ASM.AILight;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
