+++ AILZ80ASM/LineDetailExpansionItemBinaryFile.cs
@@ -1,6 +1,8 @@
-﻿using System;
-using System.IO;
+﻿using AILZ80ASM.AILight;
+using AILZ80ASM.Assembler;
 using AILZ80ASM.Exceptions;
+using System;
+using System.IO;
 
 namespace AILZ80ASM
 {
