+++ AILZ80ASM/LineDetailItemFunctionDefine.cs
@@ -1,6 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using AILZ80ASM.AILight;
+using AILZ80ASM.Assembler;
+using System;
 using System.Text.RegularExpressions;
 
 namespace AILZ80ASM
