+++ AILZ80ASM/LineDetailItemInclude.cs
@@ -1,9 +1,10 @@
-﻿using System;
+﻿using AILZ80ASM.Assembler;
+using AILZ80ASM.Exceptions;
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
-using AILZ80ASM.Exceptions;
 
 namespace AILZ80ASM
 {
