+++ AILZ80ASM/LineDetailItemMacroDefineCompatible.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using AILZ80ASM.Assembler;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
