+++ AILZ80ASM/LineDetailItemRepeat.cs
@@ -1,8 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
+using AILZ80ASM.Assembler;
+using AILZ80ASM.Exceptions;
 using System.Linq;
 using System.Text.RegularExpressions;
-using AILZ80ASM.Exceptions;
+using AILZ80ASM.AILight;
 
 namespace AILZ80ASM
 {
