+++ AILZ80ASM/LineDetailScopeItem.cs
@@ -1,4 +1,5 @@
-﻿using AILZ80ASM.Exceptions;
+﻿using AILZ80ASM.Assembler;
+using AILZ80ASM.Exceptions;
 using System;
 using System.Linq;
