+++ AILZ80ASM/LineItem.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿using AILZ80ASM.AILight;
+using AILZ80ASM.Assembler;
+using System.IO;
 
 namespace AILZ80ASM
 {
