+++ AILZ80ASM/OperationItemDataSpace.cs
@@ -1,8 +1,10 @@
-﻿using System;
+﻿using AILZ80ASM.AILight;
+using AILZ80ASM.Assembler;
+using AILZ80ASM.Exceptions;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using AILZ80ASM.Exceptions;
 
 namespace AILZ80ASM
 {
