+++ AILZ80ASM/Program.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using AILZ80ASM.Assembler;
+using AILZ80ASM.CommandLine;
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Text.Json;
-using AILZ80ASM.CommandLine;
 
 namespace AILZ80ASM
 {
